# Add stream selection early exit