`select_early_exit` allows you to create a stream that produces elements while either stream contains elements, and terminates when either stream runs out of elements.

## Previous work

https://github.com/rust-lang/futures-rs/pull/1881
https://github.com/rust-lang/futures-rs/pull/1422

## Benchmarks:

| Old | New |
| --- | --- |
| 1,065,438ns/iter | 1,159,341ns/iter |

Performance hit: +8.8%.

The `select_streams` [benchmark](https://github.com/rust-lang/futures-rs/pull/2582/files) is quite contrived, using long streams nested deeply under multiple `Select`s to produce statistically significant results. As such, I'm happy with the +8.8% performance hit, and would prefer it over a separate version of `SelectWithStrategy` just for early exit.

Side note, if you leave out the `#[inline]` over the `is_finished` function, you get `1,830,713ns/iter`, or a +71.8% performance hit.

I've tried to specialize `SelectWithStrategy` over an exit strategy trait in [this branch](https://github.com/414owen/futures-rs/tree/early-exit-streams-as-type-param), but surprisingly, got a small performance hit.

I can't say for sure what's going on, but my best guess is that the performance difference between this branch and the one linked above is due to a layout change, rather than how the logic is encoded.

It would be possible to use something like [`stabilizer`](https://github.com/ccurtsinger/stabilizer) to know for sure, but I doubt we actually want to go that deep on this.